PMM-7616 apply TabbedContent to Inventory
