+++ public/app/percona/inventory/Inventory.messages.ts
@@ -1,0 +1,7 @@
+export const Messages = {
+  tabs: {
+    services: 'Services',
+    agents: 'Agents',
+    nodes: 'Nodes',
+  },
+};
