+++ public/app/percona/inventory/Inventory.styles.ts
@@ -1,22 +1,10 @@
 import { css } from 'emotion';
 
-export const tabContent = css`
-  height: 100%;
-
-  th {
-    position: sticky;
-    top: 0;
-  }
-`;
-
-export const inventoryWrapper = css`
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  width: 100%;
-`;
-
-export const tabContentWrapper = css`
-  height: 100%;
-  overflow: hidden;
-`;
+export const getStyles = () => ({
+  inventoryWrapper: css`
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    width: 100%;
+  `,
+});
